Move claim type shared contexts to own file
Should not load here because it mean it is
always loaded (slows spec boot), and it breaks
smoke test (has not access to rspec).
